backen/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
sqlalchemy==1.4.49
pydantic==1.10.11
python-dotenv==1.0.0
httpx==0.24.1
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
httpx
typing_extensions
